chore: rename plugins folder
