+++ plugin/android/constants.js
