+++ plugin/android/withApplyAndroidCompileOptions.js
