+++ plugin/android/withApplyLineImplementation.js
