+++ plugin/android/withApplyMaven.js
