+++ plugin/android/withLineManifest.js
