+++ plugin/ios/appDelegate_template.m
