+++ plugin/ios/withAppDelegateMod.js
