+++ plugin/ios/withCustomPlist.js
